# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Order Transaction API",
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
...
        "/product": {
            "get": {
                "tags": [
                    "/product"
                ],
                "summary": "List Product",
                "description": "List Product",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "sku",
                        "in": "query",
                        "description": "Product Stock Keeping Unit",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                }
            },
            "post": {
                "tags": [
                    "/product"
                ],
                "summary": "Create new Product",
                "description": "Create new Product",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "sku",
                        "in": "formData",
                        "description": "Product Stock Keeping Unit",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Product Price",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "qty",
                        "in": "formData",
                        "description": "Product Quantity",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "description": "Product Description",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "/product"
                ],
                "summary": "View Product detail",
                "description": "View Product detail",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/product"
                ],
                "summary": "Update existing Product",
                "description": "Update existing Product",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Product Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Product Price",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "qty",
                        "in": "formData",
                        "description": "Product Quantity",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "description": "Product Description",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/product/{id}/status": {
            "get": {
                "tags": [
                    "/product"
                ],
                "summary": "Change Product Status",
                "description": "Change Product status",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "x-access-token",
            "in": "header"
        }
    },
...
}